+++ core/objects/finders/Bvh.h
@@ -4,7 +4,6 @@
 #include "objects/geometry/Box.h"
 #include "objects/geometry/Indices.h"
 #include "objects/wrappers/AutoPtr.h"
-#include <set>
 
 NAMESPACE_SPH_BEGIN
 
@@ -42,7 +41,7 @@ public:
 ///
 /// There is always either zero or two intersections. If the ray intersects the box, the function returns true
 /// and the intersection distances (near and far distance) are returned as t_min and t_max.
-bool intersectBox(const Box& box, const Ray& ray, Float& t_min, Float& t_max);
+inline bool intersectBox(const Box& box, const Ray& ray, Float& t_min, Float& t_max);
 
 
 struct BvhPrimitive {
@@ -103,7 +102,7 @@ public:
         }
         intersection.object = this;
         intersection.t = f * dot(dir2, q);
-        return true;
+        return intersection.t > 0.f;
     }
 
     INLINE Box getBBox() const {
@@ -145,7 +144,7 @@ public:
 
         intersection.object = this;
         intersection.t = deltaCos - sqrt(disc);
-        return true;
+        return intersection.t > 0.f;
     }
 
     INLINE Box getBBox() const {
@@ -172,7 +171,7 @@ public:
         if (result) {
             intersection.t = t_min;
             intersection.object = this;
-            return true;
+            return intersection.t > 0.f;
         } else {
             return false;
         }
@@ -187,24 +186,6 @@ public:
     }
 };
 
-class IBvh : public Polymorphic {
-public:
-    /// \brief Finds the closest intersection of the ray.
-    ///
-    /// Returns true if an intersection has been found.
-    virtual bool getFirstIntersection(const Ray& ray, IntersectionInfo& intersection) const = 0;
-
-    /// \brief Returns all intersections of the ray.
-    virtual void getAllIntersections(const Ray& ray, std::set<IntersectionInfo>& intersections) const = 0;
-};
-
-struct BvhNode {
-    Box box;
-    Size start;
-    Size primCnt;
-    Size rightOffset;
-};
-
 /// \brief Simple bounding volume hierarchy.
 ///
 /// Interface for finding an intersection of given ray with a set of geometric objects. Currently very
@@ -212,13 +193,20 @@ struct BvhNode {
 /// other geometric primitives are needed, either add the specialization to cpp, or move the implementation to
 /// header.
 template <typename TBvhObject>
-class Bvh : public IBvh {
+class Bvh : public Noncopyable {
 private:
     const Size leafSize;
     Size nodeCnt = 0;
     Size leafCnt = 0;
 
     Array<TBvhObject> objects;
+
+    struct BvhNode {
+        Box box;
+        Size start;
+        Size primCnt;
+        Size rightOffset;
+    };
 
     Array<BvhNode> nodes;
 
@@ -234,11 +222,14 @@ public:
     /// \brief Finds the closest intersection of the ray.
     ///
     /// Returns true if an intersection has been found.
-    virtual bool getFirstIntersection(const Ray& ray, IntersectionInfo& intersection) const override;
+    bool getFirstIntersection(const Ray& ray, IntersectionInfo& intersection) const;
 
     /// \brief Returns all intersections of the ray.
-    virtual void getAllIntersections(const Ray& ray,
-        std::set<IntersectionInfo>& intersections) const override;
+    ///
+    /// Intersections are returned via a generic output iterator.
+    /// \returns Total number of intersections.
+    template <typename TOutIter>
+    Size getAllIntersections(const Ray& ray, TOutIter iter) const;
 
     /// \brief Returns true if the ray is occluded by some geometry
     bool isOccluded(const Ray& ray) const;
@@ -252,3 +243,5 @@ private:
 };
 
 NAMESPACE_SPH_END
+
+#include "objects/finders/Bvh.inl.h"