+++ gui/renderers/VolumeRenderer.cpp
@@ -3,6 +3,7 @@
 #include "gui/objects/Camera.h"
 #include "gui/objects/Colorizer.h"
 #include "gui/renderers/FrameBuffer.h"
+#include "objects/utility/OutputIterators.h"
 
 NAMESPACE_SPH_BEGIN
 
@@ -31,7 +32,7 @@ void VolumeRenderer::initialize(const Storage& storage,
     bvh.build(std::move(spheres));
 
     for (ThreadData& data : threadData) {
-        data.data = std::set<IntersectionInfo>{};
+        data.data = RayData{};
     }
 
     shouldContinue = true;
@@ -41,10 +42,15 @@ Rgba VolumeRenderer::shade(const RenderParams& params, const CameraRay& cameraRa
     const Vector dir = getNormalized(cameraRay.target - cameraRay.origin);
     const Ray ray(cameraRay.origin, dir);
 
-    std::set<IntersectionInfo>& intersections(data.data);
-    bvh.getAllIntersections(ray, intersections);
+    RayData& rayData(data.data);
+    Array<IntersectionInfo>& intersections = rayData.intersections;
+    intersections.clear();
+    bvh.getAllIntersections(ray, backInserter(intersections));
+    /// \todo only needed if absorption is enabled
+    std::sort(intersections.begin(), intersections.end());
+
     Rgba result = this->getEnviroColor(cameraRay);
-    for (const IntersectionInfo& is : intersections) {
+    for (const IntersectionInfo& is : reverse(intersections)) {
         const BvhSphere* s = static_cast<const BvhSphere*>(is.object);
         const Size i = s->userData;
         const Vector hit = ray.origin() + ray.direction() * is.t;